requirements.txt
httpx
python-dotenv
PySide6==6.6.2
PySide6-Addons==6.6.2
websockets
deepgram-sdk
azure-cognitiveservices-speech
requests
google-auth-oauthlib
google-api-python-client
fastapi
openai
aiohttp
Pillow
pytz
timezonefinder
uvicorn
